Remove TODO
The case where the combined noise signal is an array of only 0s
and noise sources are selected can cause the program to crash.
This edge case can occur when the combination of selected noise
parameters results in too few noise  samples.

In this case, the assertions from `noise_synthesis.py` display.
I think these assertions are sufficient for prompting the user to
adjust their input parameters. This is the extent of error handling
I know for streamlit.

Issue may be revisited when working on streaming data.