Add docstring to quantization noise.
Previously, `simulate_quantization_noise` did not have a docstring.
Now a docstring is provided.
There is no physical reason that the amplitude or frequency of a
pure tone sinewave needs to be an integer. So, the `noise_amp` and
`noise_freq` arguments were changed to default floats.

This commit also touches up the docstrings for other methods.
Each method produces a series of values with a natural ordinality.
So, it made sense to update the language in the docstrings to say
a **time** series is being produced.