Update sidebar title and app body
The term "Error Coefficients" refers to the characteristic constants
that define each noise type; they are commonly found on hardware data
sheets. The term "Noise Parameters" is not the accepted jargon. So,
the sidebar title was changed to "Error Coefficients" to use the more
widely accepted term.

Since the sidebar title was changed, the main text of the app body that
referred to this section of the sidebar was updated to use "Error Coefficients".

Some minor changes in the appy body were made. Specifically, some
language was changed to make it explicit that only a single gyroscope
is being simulated. This is necessary since an IMU is actually made
by combining 3 acceleometers and gyroscopes.

Multi-line plots are on the roadmap and the language should be
updated after that feature is implemented.