Lower number input to handle *nano* level values
The previous limit was an arbitrary choice made by looking up a
random IMU data sheet and assuming the listed values were the absolute
minima. One idea was to include a `units` dropdown for the inputs
that would specify the magnitude of the input.

Since the input minimum is a function argument, it can be set to
nearly any value. So, there is no need for a dropdown menu of units.
However, in the example given with this dropdown menu, the lowest
unit was nano. So, I am assuming 10e-9 is the lowest magnitude.
The input minima have been changed according to this new assumption.

Formatting was also changed in this commit. Since the inputs can
take on nano level precision, the default behavior is to use
lowercase scientific notation when possible. Else, to list
each significant digit.