Calculate time-axis values from sim time and fs
Previously, the default behavior was to use the index of the time
series as the x axis for the time series plot. But, the simulation
length and sampling frequency determine these timestamps.

Now, the timestamps run from 0 to `sim_time`. The number of timestamps
is calculated as

`int(sim_time*fs)`

which corresponds to N seconds * M samples/second