Add quantization noise coefficient parameter
There must be an associated coefficient for quantization noise.
The algorithm used to generate quantization noise follows from
calculating `q`, however `q` itself is not the coefficient.

The value `q` depends on 2 inputs:  `K` and `E`. Clearly, `E` is
a function of the simulation length and sampling frequency and `K`
is a constant. Previously, we had `K=1`. It's clear now that the
choice `K=1` was arbitrary to make the demo code work.

Since `K` is a constant in the demo code and there are no constant
variables, I conclude that `K` must be the error coefficient for
quantization noise.

Then, this commit changes the function signature to require the
error coefficient is input.